+++ mediark/infrastructure/core/configuration/production_config.py
@@ -28,9 +28,9 @@ class ProductionConfig(DevelopmentConfig):
             "JwtSupplier": {
                 "method": "jwt_supplier"
             },
-            "Authenticate": {
-                "method": "middleware_authenticate"
-            },
+            # "Authenticate": {
+            #     "method": "middleware_authenticate"
+            # },
             "ImageRepository": {
                 "method": "shelve_image_repository",
             },
