+++ mediark/infrastructure/web/api.py
@@ -17,6 +17,7 @@ def create_api(app: Flask, resolver: Injectark) -> None:
     app.add_url_rule("/", view_func=root_view)
 
     # Middleware
+    print('resolver>>>>>>', type(resolver))
     authenticate = resolver['Authenticate']
 
     # Images Resource
